plgx-esp-ui/requirements/prod.txt
alembic==1.4.2          # via flask-migrate
amqp==2.6.0          # via celery -> kombu
aniso8601==8.0.0          # via flask-restplus
antlr4-python3-runtime==4.8          # via stix2 -> stix2-patterns
attrs==19.3.0          # via flask-restplus -> jsonschema

bcrypt==3.1.7          # via flask-bcrypt
billiard==3.5.0.5          # via celery
blinker==1.4          # via flask-debugtoolbar

celery==4.1.1
certifi==2020.6.20          # via requests-oauthlib, requests
cffi==1.14.0          # via flask-bcrypt -> bcrypt
click==7.1.2          # via flask

Flask==1.0
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.8
Flask-DebugToolbar==0.11.0
Flask-HTTPAuth==4.1.0
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.5.3
Flask-Script==2.0.6
Flask-SocketIO==4.3.1
Flask-Sockets==0.2.1
Flask-SQLAlchemy==2.4.4          # via flask-migrate
Flask-SSLify==0.1.5
Flask-wtf==0.14.3          # via flask, flask-ldap3-login

flask-ldap3-login==0.9.16
flask-restplus==0.13.0

gevent==1.4.0          # via flask-sockets
gevent-websocket==0.10.1          # via flask-sockets
greenlet==0.4.16          # via flask-sockets -> gevent
gunicorn==20.0.4

idna==2.10          # via requests-oauthlib, requests
importlib-metadata==1.7.0          # via flask-restplus -> jsonschema
itsdangerous==1.1.0          # via flask

Jinja2==2.11.2          # via flask
jsonschema==3.2.0          # via flask-restplus
jsmin==2.2.2

kombu==4.6.11          # via celery

ldap3==2.7          # via flask-ldap3-login

Mako==1.1.3          # via flask-migrate -> alembic
MarkupSafe==1.1.1          # via alembic -> mako, flask -> jinja2
more-itertools==8.4.0          # via pyamqp -> pytest

oauthlib==3.1.0          # via requests-oauthlib

packaging==20.4          # via pyamqp -> pytest
pika==0.13.0          # via pyamqp
pluggy==0.13.1          # via pyamqp -> pytest
psycopg2==2.8
py==1.9.0          # via pyamqp -> pytest
PyAMQP==0.1.0.7
pyasn1==0.4.8          # via flask-ldap3 -> ldap3
pycparser==2.20          # via flask-bcrypt -> bcrypt -> cffi
PyJWT==1.7.1
pyparsing==2.4.7          # via pyamqp -> pytest -> packaging
pyrsistent==0.16.0          # via flask-restplus -> jsonschema
pytest==5.4.3          # via pyamqp
python-dateutil==2.8.1          # via flask-migrate -> alembic
python-editor==1.0.4          # via flask-migrate -> alembic
python-engineio==3.13.1          # via flask-socketio -> python-socketio
python-socketio==4.6.0          # via flask-socketio
pytz==2020.1          # via celery

raven==6.10.0
redis==3.5.3
requests==2.24.0          # via requests-oauthlib
requests-oauthlib==1.3.0

setuptools==41.0.1          # via gunicorn, flask-restplus -> jsonschema
simplejson==3.17.2          # via stix2
six==1.15.0          # via flask-bcrypt -> bcrypt, flask-migrate -> alembic -> python-dateutil
socketIO-client3==0.8.0
SQLAlchemy==1.3.17          # via flask-migrate -> alembic

taxii2-client==0.5.0          # keep it before the stix installation always
stix2==1.1.2
stix2-patterns==1.3.1          # via stix2

unicodecsv==0.14.1
urllib3==1.25.9          # via requests-oauthlib, requests

vine==1.3.0          # via celery -> kombu -> ampq

wcwidth==0.2.5          # via pyamqp -> pytest
websocket-client==0.57.0          # via socketio-client
Werkzeug==0.15.3          # via flask
wheel==0.33.1
WTForms==2.3.1          # via flask-wtf

zipp==3.1.0          # via flask-restplus -> jsonschema -> importlib-metadata
